requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-groq
faiss-cpu
sentence-transformers
huggingface-hub
numpy
pandas
python-dotenv
fastapi
uvicorn
